// src/forms/enums/forms-messages.enum.ts
export enum FormSuccessMessages {
  CREATED = "فرم ایجاد شد",
  FOUND = "فرم پیدا شد",
  LISTED = "لیست فرم ها دریافت شد",
  UPDATED = "فرم با موفقیت بروز رسانی شد",
  DELETED = "فرم با موفقیت حذف شد",
}

export enum FormErrorMessages {
  DUPLICATE = "فرم تکراری است",
  NOT_FOUND = "فرم پیدا نشد",
  FORM_NOT_FOUND = "برای نوع فرم باید شناسه فرم مشخص باشد",
}